refactor: 리뷰 반영 리팩토링
