+++ src/main/java/step3/domain/WinningInfo.java
@@ -1,4 +1,4 @@
-package step2.domain;
+package step3.domain;
 
 public class WinningInfo {
     private final int matchCount;
