+++ src/main/java/step3/domain/WinningLotto.java
@@ -1,4 +1,4 @@
-package step2.domain;
+package step3.domain;
 
 import java.util.LinkedHashMap;
 import java.util.List;
