+++ src/main/java/step3/util/GeneratorUtil.java
@@ -1,4 +1,4 @@
-package step2.util;
+package step3.util;
 
 import java.util.ArrayList;
 import java.util.Collections;
