+++ src/main/java/step3/view/InputView.java
@@ -1,4 +1,4 @@
-package step2.view;
+package step3.view;
 
 import java.util.Arrays;
 import java.util.List;
