+++ src/test/java/step3/domain/LottoTest.java
@@ -1,4 +1,4 @@
-package step2.domain;
+package step3.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
