+++ src/test/java/step3/domain/WinningLottoTest.java
@@ -1,26 +1,26 @@
-package step2.domain;
+package step3.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class WinningLottoTest {
 
+    Lotto lottoFixture(Integer... numbers){
+        return new Lotto(Stream.of(numbers)
+                    .map(LottoNumber::of)
+                    .collect(Collectors.toList()));
+    }
+
     @Test
     @DisplayName("당첨번호를 입력으로 넣으면 당첨금 2_000_000_000원이다.")
     void winningLottoTest() {
         // given
-        Lotto winningNumbersLotto = new Lotto(List.of(
-                LottoNumber.of(1),
-                LottoNumber.of(2),
-                LottoNumber.of(3),
-                LottoNumber.of(4),
-                LottoNumber.of(5),
-                LottoNumber.of(6)
-        ));
+        Lotto winningNumbersLotto = lottoFixture(1, 2, 3, 4, 5, 6);
 
         Lottos lottos = new Lottos();
         lottos.addLotto(winningNumbersLotto);
